+++ api/server/router/volume/volume_routes.go
@@ -6,7 +6,9 @@ import (
 
 	"github.com/docker/docker/api/server/httputils"
 	"github.com/docker/docker/api/types/filters"
+	"github.com/docker/docker/api/types/versions"
 	volumetypes "github.com/docker/docker/api/types/volume"
+	volumestore "github.com/docker/docker/volume/store"
 	"golang.org/x/net/context"
 )
 
@@ -49,6 +51,13 @@ func (v *volumeRouter) postVolumesCreate(ctx context.Context, w http.ResponseWri
 	}
 
 	volume, err := v.backend.VolumeCreate(req.Name, req.Driver, req.DriverOpts, req.Labels)
+	if volumestore.IsAlreadyExists(err) {
+		version := httputils.VersionFromContext(ctx)
+		if versions.GreaterThanOrEqualTo(version, "1.29") {
+			return httputils.WriteJSON(w, http.StatusNotModified, volume)
+		}
+		return httputils.WriteJSON(w, http.StatusCreated, volume)
+	}
 	if err != nil {
 		return err
 	}